requirements.txt
langchain
langchain-groq
langchain-community
langchain-huggingface
langgraph
langgraph-checkpoint-sqlite
faiss-cpu
sentence-transformers
torch
pydantic
python-dotenv
streamlit
tiktoken
aiosqlite
